# Add new optimizer state `row_counter` for Adam [Frontend]